Rearraged project to its final structure
