+++ rocket-chatter-ingestion-server/src/process/prepare/processor/core/treeNode.ts
@@ -1,4 +1,4 @@
-export class DBNode {
+export class TreeNode {
 	name: string = ""
 	type: string = ""
 	body: string = ""
