+++ rocket-chatter-ingestion-server/src/process/prepare/processor/syntax/functions.ts
@@ -1,13 +1,13 @@
 import { namedTypes } from "ast-types"
 import { print } from "recast"
+import { TreeNode } from "../core/treeNode"
+import { TypeAnnotation } from "../core/typeAnnotation"
+import { TypeArgument } from "../core/typeArgument"
 import { Classes } from "./classes"
-import { DBNode } from "./fundamental/dbNode"
-import { TypeAnnotation } from "./fundamental/typeAnnotation"
-import { TypeArgument } from "./fundamental/typeArgument"
 
 export namespace Functions {
 	export function Handle(n: namedTypes.FunctionDeclaration) {
-		const node = new DBNode(
+		const node = new TreeNode(
 			n.id?.name.toString() ?? "",
 			"function",
 			n.body.body.map((e) => print(e).code).join("\n")
